Update a few missed areas, clean up

+++ tests/algorithms/scaling/test_Ih_table.py
@@ -150,7 +150,7 @@ def test_IhTableblock_onedataset(large_reflection_table, test_sg):
     n_groups = len(set(asu_indices))
     n_refl = large_reflection_table.size()
     block = IhTableBlock(n_groups=n_groups, n_refl=n_refl)
-    group_ids = np.array([0, 1, 0, 2, 3, 4, 4])
+    group_ids = np.array([0, 1, 0, 2, 3, 4, 4], dtype=np.uint)
 
     df = refl_table_to_df(large_reflection_table)
 
@@ -280,12 +280,15 @@ def test_IhTableblock_twodatasets(large_reflection_table, test_sg):
 
     block.add_data(
         0,
-        np.array([0, 1, 3, 4, 4]),
+        np.array([0, 1, 3, 4, 4], dtype=np.uint),
         refl_table_to_df(dataset_1),
         dataset_1["miller_index"],
     )
     block.add_data(
-        1, np.array([0, 2]), refl_table_to_df(dataset_2), dataset_2["miller_index"]
+        1,
+        np.array([0, 2], dtype=np.uint),
+        refl_table_to_df(dataset_2),
+        dataset_2["miller_index"],
     )
 
     assert list(block.block_selections[0]) == [0, 1, 2, 3, 4]
